Added base url to env files

+++ pages/api/products/[id].js
@@ -6,9 +6,12 @@ export default async function handleProduct(req, res) {
   const {
     method,
     query: { id },
+    cookies,
   } = req;
 
-  dbConnect();
+  const token = cookies.token;
+
+  await dbConnect();
 
   if (method === "GET") {
     try {
@@ -20,6 +23,10 @@ export default async function handleProduct(req, res) {
   }
 
   if (method === "PUT") {
+    if (!token || token !== process.env.COOKIE_TOKEN) {
+      return res.status(401).json("Not authenticated.");
+    }
+
     try {
       const product = await Product.create(req.body);
       res.status(201).json(product);
@@ -30,6 +37,10 @@ export default async function handleProduct(req, res) {
   }
 
   if (method === "DELETE") {
+    if (!token || token !== process.env.COOKIE_TOKEN) {
+      return res.status(401).json("Not authenticated.");
+    }
+
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json("The product has been deleted");
